% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_add_RLumTeam.R
\name{module_add_RLumTeam}
\alias{module_add_RLumTeam}
\title{Module: add RLum-Team and update version timestamp}
\usage{
module_add_RLumTeam()
}
\description{
If the package was supported by the RLumTeam (the standard) the RLum-Team is added
as author to all functions.
}
\section{Function version}{
 0.1.0
}

\author{
Sebastian Kreutzer, Institute of Geography, Heidelberg University (Germany)
, RLum Developer Team}
