% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_build_documentation.R
\name{module_build_documentation}
\alias{module_build_documentation}
\title{Module: Build Documentation}
\usage{
module_build_documentation()
}
\description{
Create documentation using roxygen2
}
\section{Function version}{
 0.1.0
}

\author{
Sebastian Kreutzer, Institute of Geography, Heidelberg University (Germany)
, RLum Developer Team} 

\section{How to cite}{
Kreutzer, S., 2024. module_build_documentation(): Module: Build Documentation. Function version 0.1.0. In: Kreutzer, S., Burow, C., 2024. RLumBuild: RLum Universe Package Building. R package version 0.1.8.9000-5. https://github.com/R-Lum/RLumBuild
}

