% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_check_ReverseDependencies.R
\name{module_check_ReverseDependencies}
\alias{module_check_ReverseDependencies}
\title{Check Reverse Dependencies}
\usage{
module_check_ReverseDependencies()
}
\description{
Check the reverse dependencies of the package
}
\section{Function version}{
 0.1.0
}

\author{
Sebastian Kreutzer, Institute of Geography, Heidelberg University (Germany)
, RLum Developer Team}
