% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_compactVignette.R
\name{module_compactVignette}
\alias{module_compactVignette}
\title{Module: compact_Vignette}
\usage{
module_compactVignette()
}
\description{
Try to further downsize the vignette, so far available
}
\section{Function version}{
 0.1.0
}

\author{
Sebastian Kreutzer, Institute of Geography, Heidelberg University (Germany)
, RLum Developer Team} 

\section{How to cite}{
Kreutzer, S., 2025. module_compactVignette(): Module: compact_Vignette. Function version 0.1.0. In: Kreutzer, S., Burow, C., 2025. RLumBuild: RLum Universe Package Building. R package version 0.1.10.9000-18. https://github.com/R-Lum/RLumBuild
}

