% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_compile_Attributes.R
\name{module_compile_Attributes}
\alias{module_compile_Attributes}
\title{Module: Compile attributes}
\usage{
module_compile_Attributes()
}
\description{
Compile C++ attributes if such code is part of the package
}
\section{Function version}{
 0.1.0
}

\author{
Sebastian Kreutzer, Institute of Geography, Heidelberg University (Germany)
, RLum Developer Team}
