% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_knit_README.R
\name{module_knit_README}
\alias{module_knit_README}
\title{Module: Knit README}
\usage{
module_knit_README()
}
\description{
Create README.md from README.Rmd
}
\section{Function version}{
 0.1.0
}

\author{
Sebastian Kreutzer, Institute of Geography, Heidelberg University (Germany)
, RLum Developer Team}
