% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_register_EntryPoints.R
\name{module_register_EntryPoints}
\alias{module_register_EntryPoints}
\title{Module: set_VersionNumber}
\usage{
module_register_EntryPoints(pkg_name)
}
\arguments{
\item{pkg_name}{\link{character} the name of the package}
}
\description{
Module: set_VersionNumber
}
