% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_set_VersionNumber.R
\name{module_set_VersionNumber}
\alias{module_set_VersionNumber}
\title{Module: set_VersionNumber}
\usage{
module_set_VersionNumber()
}
\description{
Automatically set version number in DESCRIPTION file if the pattern '.9000' is found.
}
\details{
If the package is under development it is somehow meaningful to easily distinguish
between different build versions. In the R community the pattern `.9000` in a version number is
used to declare a package under development. If such pattern is found, the module
either adds `-1` or if already their, increase such number by `1` every time the package
is built.

For release versions the patter `.9000` is removed from the DESCRIPTION file and nothing
is done if the module is run.
}
\section{Function version}{
 0.1.0
}

\author{
Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom)
, RLum Developer Team} 

\section{How to cite}{
Kreutzer, S., 2020. module_set_VersionNumber(): Module: set_VersionNumber. Function version 0.1.0. In: Kreutzer, S., Burow, C., 2020. RLumBuild: RLum Universe Package Building. R package version 0.1.5.9000-6. https://github.com/R-Lum/RLumBuild
}

