% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_update_datalist.R
\name{module_update_datalist}
\alias{module_update_datalist}
\title{Update datalist}
\usage{
module_update_datalist()
}
\description{
Updates the file `data/datalist` if available. The standard CRAN functionality
only works for data > 1 MB. For whatever reason.
}
\section{Function version}{
 0.1.0
}

\author{
Sebastian Kreutzer, Institute of Geography, Heidelberg University (Germany)
, RLum Developer Team}
