% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_update_zenodoJSON.R
\name{module_update_zenodoJSON}
\alias{module_update_zenodoJSON}
\title{Module: Update .zenodo.json file}
\usage{
module_update_zenodoJSON()
}
\description{
Automatically update the .zenodo.json file
}
\details{
A small helper to make sure that we have not to take care
update `version`, `title` and `description` in the `zenodo.json` file
}
\section{Function version}{
 0.1.0
}

\author{
Sebastian Kreutzer, Institute of Geography, Heidelberg University (Germany)
, RLum Developer Team}
