% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_verify_ExampleTimings.R
\name{module_verify_ExampleTimings}
\alias{module_verify_ExampleTimings}
\title{Verify Example Timings}
\usage{
module_verify_ExampleTimings()
}
\description{
Check and display the timing results of the example checks to avoid too long
example runs.
}
\section{Function version}{
 0.1.0
}

\author{
Sebastian Kreutzer, Institute of Geography, Heidelberg University (Germany)
, RLum Developer Team}
