% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_write_BibTeX.R
\name{module_write_BibTeX}
\alias{module_write_BibTeX}
\title{Create Package BibTeX file}
\usage{
module_write_BibTeX()
}
\description{
A BibTeX file to be used to cite the package is automatically created.
}
\section{Function version}{
 0.1.0
}

\author{
Sebastian Kreutzer, Institute of Geography, Heidelberg University (Germany)
, RLum Developer Team}
