% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_write_codemetar.R
\name{module_write_codemetar}
\alias{module_write_codemetar}
\title{Create Codemetar JSON-LD file}
\usage{
module_write_codemetar()
}
\description{
Add JSON-LD file file to make the package more tracable
}
\section{Function version}{
 0.1.0
}

\seealso{
[codemetar::create_codemeta]
}
\author{
Sebastian Kreutzer, Institute of Geography, Heidelberg University (Germany)
, RLum Developer Team}
