% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_write_FunctionList.R
\name{module_write_FunctionList}
\alias{module_write_FunctionList}
\title{Write Function List}
\usage{
module_write_FunctionList()
}
\description{
Create a HTML and markdown list of all major functions in the package
}
\section{Function version}{
 0.1.0
}

\author{
Sebastian Kreutzer, IRAMAT-CRP2A, UMR 5060, CNRS - Université Bordeaux Montaigne (France)
} 

\section{How to cite}{
Kreutzer, S. (2019). module_write_FunctionList(): Write Function List. Function version 0.1.0. In: Kreutzer, S., Burow, C. (2019). RLumBuild: RLum Universe Package BuildingR package version 0.1.0.9000-27. https://CRAN.R-project.org/package=RLumBuild
}

