% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_write_PDF_manual.R
\name{module_write_PDF_manual}
\alias{module_write_PDF_manual}
\title{Create PDF Manual}
\usage{
module_write_PDF_manual()
}
\description{
Build PDF Manual tar.gz is available
}
\section{Function version}{
 0.1.0
}

\seealso{
[devtools::build_manual]
}
\author{
Sebastian Kreutzer, Institute of Geography, Heidelberg University (Germany)
, RLum Developer Team} 

\section{How to cite}{
Kreutzer, S., 2024. module_write_PDF_manual(): Create PDF Manual. Function version 0.1.0. In: Kreutzer, S., Burow, C., 2024. RLumBuild: RLum Universe Package Building. R package version 0.1.10.9000-11. https://github.com/R-Lum/RLumBuild
}

