% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLumBuild-package.R
\docType{package}
\name{RLumBuild-package}
\alias{RLumBuild-package}
\alias{RLumBuild}
\title{RLum Universe Package Building
\if{html}{
\figure{RLumBuild-logo.svg}{options: width="50px"}\cr
}}
\description{
Build packages from the RLum-universe.
A collection of tools and scripts to unify the building of packages.
}
\details{
\strong{Funding}

In 2019, the work of Sebastian Kreutzer as maintainer of the package was supported
by the LabEx LaScArBxSK (ANR - n. ANR-10-LABX-52).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/R-Lum/RLumBuild}
  \item Report bugs at \url{https://github.com/R-Lum/RLumBuild/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Kreutzer \email{sebastian.kreutzer@uni-heidelberg.de} (\href{https://orcid.org/0000-0002-0734-2199}{ORCID})

Authors:
\itemize{
  \item Sebastian Kreutzer \email{sebastian.kreutzer@uni-heidelberg.de} (\href{https://orcid.org/0000-0002-0734-2199}{ORCID})
  \item Christoph Burow (\href{https://orcid.org/0000-0002-5023-4046}{ORCID})
}

}
\keyword{package}
