% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_package.R
\name{build_package}
\alias{build_package}
\title{Check and Build Package}
\usage{
build_package(exclude = NULL)
}
\arguments{
\item{exclude}{\link{character} (optional): names of build modules you want to exclude}
}
\description{
Check and Build Package
}
\examples{

print("test")

}
