% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{helpers}
\alias{helpers}
\alias{.running}
\alias{.success}
\alias{.failure}
\alias{.neutral}
\alias{.shut_up}
\alias{.run_module}
\alias{.get_pkg_name}
\alias{.get_pkg_version}
\title{RLumBuild Helper Functions}
\usage{
.running(text = "")

.success(text = "")

.failure(text = "")

.neutral(text = "")

.shut_up(expr)

.run_module(text = "", f = NULL, shut_up = TRUE)

.get_pkg_name()

.get_pkg_version()
}
\arguments{
\item{text}{\link{character} (with default): text to display in the terminal}

\item{expr}{\link{expression} (\strong{required}): R expression to be excecuted in silence, the results is piped
through and the output muted (messages, warnings and usual terminal prints). Errors are returned.}

\item{shut_up}{\link{logical} (with default): enables/disables shut_up mode}
}
\description{
Helper functions which are used internally to visualise the console output.
Those functions are not in the overivew, but are, however, documented
}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, UMR 5060, CNRS - Université Bordeaux Montaigne (France)
} 

\section{How to cite}{
Kreutzer, S. (2019). helpers(): RLumBuild Helper Functions. In: Kreutzer, S., Burow, C. (2019). RLumBuild: RLum Universe Package BuildingR package version 0.1.1.9000-2. https://CRAN.R-project.org/package=RLumBuild
}

\keyword{internal}
