% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_add_HowToCite.R
\name{module_add_HowToCite}
\alias{module_add_HowToCite}
\title{Add How to Cite Section}
\usage{
module_add_HowToCite()
}
\description{
Adds a function tailored section 'How to Cite' to each manual page as long as
author names and a section 'Function version' exists.
}
\details{
\strong{Output example}

Burow, C., Kreutzer, S. (2019). module_add_HowToCite(): Add How to Cite Section. Function version 0.1.0.
In: Kreutzer, S., Burow, C. (2019). RLumBuild: RLum Universe Package BuildingR
package version 0.1.1.9000-5. https://github.com/R-Lum/RLumBuild
}
\section{Function version}{
 0.3.0
}

\author{
Christoph Burow (Germany), Sebastian Kreutzer, IRAMAT-CRP2A, UMR 5060, CNRS - Université Bordeaux Montaigne (Frange)
, RLum Developer Team} 

\section{How to cite}{
Burow, C., Kreutzer, S., 2019. module_add_HowToCite(): Add How to Cite Section. Function version 0.3.0. In: Kreutzer, S., Burow, C., 2019. RLumBuild: RLum Universe Package BuildingR package version 0.1.4.9000-1. https://github.com/R-Lum/RLumBuild
}

