% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_add_RLumTeam.R
\name{module_add_RLumTeam}
\alias{module_add_RLumTeam}
\title{Module: add RLum-Team and update version timestamp}
\usage{
module_add_RLumTeam()
}
\description{
If the package was supported by the RLumTeam (the standard) the RLum-Team is added
as author to all functions.
}
\section{Function version}{
 0.1.0
}

\author{
Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom)
, RLum Developer Team} 

\section{How to cite}{
Kreutzer, S., 2020. module_add_RLumTeam(): Module: add RLum-Team and update version timestamp. Function version 0.1.0. In: Kreutzer, S., Burow, C., 2020. RLumBuild: RLum Universe Package Building. R package version 0.1.5. https://github.com/R-Lum/RLumBuild
}

