% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_add_RLumTeam.R
\name{module_add_RLumTeam}
\alias{module_add_RLumTeam}
\title{Module: add RLum-Team and update version timestamp}
\usage{
module_add_RLumTeam()
}
\description{
Module: add RLum-Team and update version timestamp
}
\section{Function version}{
 0.1.0
}

\author{
Sebastian Kreutzer
} 

\section{How to cite}{
Kreutzer, S. (2019). module_add_RLumTeam(): Module: add RLum-Team and update version timestamp. Function version 0.1.0. In: Kreutzer, S., Burow, C. (2019). RLumBuild: RLum Universe Package BuildingR package version 0.1.0.9000-47. https://CRAN.R-project.org/package=RLumBuild
}

