% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_build_documentation.R
\name{module_build_documentation}
\alias{module_build_documentation}
\title{Module: Build Documentation}
\usage{
module_build_documentation()
}
\description{
Create documentation using roxygen2
}
\section{Function version}{
 0.1.0
}

\author{
Sebastian Kreutzer, IRAMAT-CRP2A, UMR 5060, CNRS - Université Bordeaux Montaigne (France)
} 

\section{How to cite}{
Kreutzer, S. (2019). module_build_documentation(): Module: Build Documentation. Function version 0.1.0. In: Kreutzer, S., Burow, C. (2019). RLumBuild: RLum Universe Package BuildingR package version 0.1.0.9000-116. https://CRAN.R-project.org/package=RLumBuild
}

