% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_check_ReverseDependencies.R
\name{module_check_ReverseDependencies}
\alias{module_check_ReverseDependencies}
\title{Check Reverse Dependencies}
\usage{
module_check_ReverseDependencies()
}
\description{
Check the reverse dependencies of the package
}
\section{Function version}{
 0.1.0
}

\author{
Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom)
, RLum Developer Team} 

\section{How to cite}{
Kreutzer, S., 2020. module_check_ReverseDependencies(): Check Reverse Dependencies. Function version 0.1.0. In: Kreutzer, S., Burow, C., 2020. RLumBuild: RLum Universe Package Building. R package version 0.1.5.9000-4. https://github.com/R-Lum/RLumBuild
}

