% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_check_ReverseDependencies.R
\name{module_check_ReverseDependencies}
\alias{module_check_ReverseDependencies}
\title{Check Reverse Dependencies}
\usage{
module_check_ReverseDependencies()
}
\description{
Check the reverse dependencies of the package
}
\section{Function version}{
 0.1.0
}

\author{
Sebastian Kreutzer, IRAMAT-CRP2A, UMR 5060, CNRS - Université Bordeaux Montaigne (France)
, RLum Developer Team} 

\section{How to cite}{
Kreutzer, S. (2019). module_check_ReverseDependencies(): Check Reverse Dependencies. Function version 0.1.0. In: Kreutzer, S., Burow, C. (2019). RLumBuild: RLum Universe Package BuildingR package version 0.1.2.9000-2. https://github.com/R-Lum/RLumBuild
}

