% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_compactVignette.R
\name{module_compactVignette}
\alias{module_compactVignette}
\title{Module: compact_Vignette}
\usage{
module_compactVignette()
}
\description{
Try to further downsize the vignette, so far available
}
\section{Function version}{
 0.1.0
}

\author{
Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom)
, RLum Developer Team} 

\section{How to cite}{
Kreutzer, S., 2020. module_compactVignette(): Module: compact_Vignette. Function version 0.1.0. In: Kreutzer, S., Burow, C., 2020. RLumBuild: RLum Universe Package Building. R package version 0.1.5.9000-4. https://github.com/R-Lum/RLumBuild
}

