% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_compile_Attributes.R
\name{module_compile_Attributes}
\alias{module_compile_Attributes}
\title{Module: Compile attributes}
\usage{
module_compile_Attributes()
}
\description{
Compile C++ attributes if such code is part of the package
}
\section{Function version}{
 0.1.0
}

\author{
Sebastian Kreutzer, IRAMAT-CRP2A, UMR 5060, CNRS - Université Bordeaux Montaigne (France)
, RLum Developer Team} 

\section{How to cite}{
Kreutzer, S., 2019. module_compile_Attributes(): Module: Compile attributes. Function version 0.1.0. In: Kreutzer, S., Burow, C., 2019. RLumBuild: RLum Universe Package BuildingR package version 0.1.4.9000-22. https://github.com/R-Lum/RLumBuild
}

