% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_create_DependencyGraph.R
\name{module_create_DependencyGraph}
\alias{module_create_DependencyGraph}
\title{Create Dependency Graph}
\usage{
module_create_DependencyGraph(
  Rversion = paste(R.version[c("major", "minor")], collapse = ".")
)
}
\arguments{
\item{Rversion}{[character] (with default): R version for which the graph is created}
}
\description{
Create package dependency graph
}
\details{
The script based on:
}
\note{
Module currently not acctivated
}
\section{Function version}{
 0.1.0
}

\author{
Sebastian Kreutzer, Institute of Geography, Heidelberg University (Germany)
, RLum Developer Team}
