% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_knit_NEWS.R
\name{module_knit_NEWS}
\alias{module_knit_NEWS}
\title{Module: Knit NEWS}
\usage{
module_knit_NEWS()
}
\description{
Create NEWS.md from NEWS.Rmd
}
\section{Function version}{
 0.1.0
}

\author{
Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom)
, RLum Developer Team} 

\section{How to cite}{
Kreutzer, S., 2020. module_knit_NEWS(): Module: Knit NEWS. Function version 0.1.0. In: Kreutzer, S., Burow, C., 2020. RLumBuild: RLum Universe Package Building. R package version 0.1.6.9000-17. https://github.com/R-Lum/RLumBuild
}

