% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_knit_NEWS.R
\name{module_knit_NEWS}
\alias{module_knit_NEWS}
\title{Module: Knit NEWS}
\usage{
module_knit_NEWS()
}
\description{
Create NEWS.md from NEWS.Rmd
}
\section{Function version}{
 0.1.0
}

\author{
Sebastian Kreutzer, Institute of Geography, Heidelberg University (Germany)
, RLum Developer Team}
