% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_knit_README.R
\name{module_knit_README}
\alias{module_knit_README}
\title{Module: Knit README}
\usage{
module_knit_README()
}
\description{
Create README.md from README.Rmd
}
\section{Function version}{
 0.1.0
}

\author{
Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom)
, RLum Developer Team} 

\section{How to cite}{
Kreutzer, S., 2020. module_knit_README(): Module: Knit README. Function version 0.1.0. In: Kreutzer, S., Burow, C., 2020. RLumBuild: RLum Universe Package Building. R package version 0.1.5.9000-6. https://github.com/R-Lum/RLumBuild
}

