% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_update_datalist.R
\name{module_update_datalist}
\alias{module_update_datalist}
\title{Update datalist}
\usage{
module_update_datalist()
}
\description{
Updates the file `data/datalist` if available. The standard CRAN functionality
only works for data > 1 MB. For whatever reason.
}
\section{Function version}{
 0.1.0
}

\author{
Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom)
, RLum Developer Team} 

\section{How to cite}{
Kreutzer, S., 2021. module_update_datalist(): Update datalist. Function version 0.1.0. In: Kreutzer, S., Burow, C., 2021. RLumBuild: RLum Universe Package Building. R package version 0.1.6.9000-20. https://github.com/R-Lum/RLumBuild
}

