% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_update_zenodoJSON.R
\name{module_update_zenodoJSON}
\alias{module_update_zenodoJSON}
\title{Module: Update .zenodo.json file}
\usage{
module_update_zenodoJSON()
}
\description{
Automatically update the .zenodo.json file
}
\details{
A small helper to keep the `zenodo.json` file up to date. This takes care
of the `version`, `title` and `description` fields. It the pacakge uses the
`Authors@R` field, the funcion will warn if the author list needs to be
updated manually (it cannot be done automatically because we do not record
the affiliation in the DESCRIPTION file).
}
\section{Function version}{
 0.1.1
}

\author{
Sebastian Kreutzer, Institute of Geography, Heidelberg University (Germany)\cr
Marco Colombo, Institute of Geography, Heidelberg University (Germany)
, RLum Developer Team} 

\section{How to cite}{
Kreutzer, S., 2025. module_update_zenodoJSON(): Module: Update .zenodo.json file. Function version 0.1.0. In: Kreutzer, S., Burow, C., 2025. RLumBuild: RLum Universe Package Building. R package version 0.1.10. https://github.com/R-Lum/RLumBuild
}

