% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_verify_ExampleTimings.R
\name{module_verify_ExampleTimings}
\alias{module_verify_ExampleTimings}
\title{Verify Example Timings}
\usage{
module_verify_ExampleTimings()
}
\description{
Check and display the timing results of the example checks to avoid too long
example runs.
}
\section{Function version}{
 0.1.0
}

\author{
Sebastian Kreutzer, Institute of Geography, Heidelberg University (Germany)
, RLum Developer Team} 

\section{How to cite}{
Kreutzer, S., 2024. module_verify_ExampleTimings(): Verify Example Timings. Function version 0.1.0. In: Kreutzer, S., Burow, C., 2024. RLumBuild: RLum Universe Package Building. R package version 0.1.8.9000-5. https://github.com/R-Lum/RLumBuild
}

