% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_verify_ExampleTimings.R
\name{module_verify_ExampleTimings}
\alias{module_verify_ExampleTimings}
\title{Verify Example Timings}
\usage{
module_verify_ExampleTimings()
}
\description{
Check and display the timing results of the example checks to avoid too long
example runs.
}
\section{Function version}{
 0.1.0
}

\author{
Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom)
, RLum Developer Team} 

\section{How to cite}{
Kreutzer, S., 2020. module_verify_ExampleTimings(): Verify Example Timings. Function version 0.1.0. In: Kreutzer, S., Burow, C., 2020. RLumBuild: RLum Universe Package Building. R package version 0.1.5.9000-6. https://github.com/R-Lum/RLumBuild
}

