% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_verify_ExampleTimings.R
\name{module_verify_ExampleTimings}
\alias{module_verify_ExampleTimings}
\title{Verify Example Timings}
\usage{
module_verify_ExampleTimings()
}
\description{
Check and display the timing results of the example checks to avoid too long
example runs.
}
\section{Function version}{
 0.1.0
}

\author{
Sebastian Kreutzer, IRAMAT-CRP2A, UMR 5060, CNRS - Université Bordeaux Montaigne (France)
, RLum Developer Team} 

\section{How to cite}{
Kreutzer, S. (2019). module_verify_ExampleTimings(): Verify Example Timings. Function version 0.1.0. In: Kreutzer, S., Burow, C. (2019). RLumBuild: RLum Universe Package BuildingR package version 0.1.2.9000-2. https://github.com/R-Lum/RLumBuild
}

