% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_verify_ExampleTimings.R
\name{module_verify_ExampleTimings}
\alias{module_verify_ExampleTimings}
\title{Veriyf Example Timings}
\usage{
module_verify_ExampleTimings()
}
\description{
Veriyf Example Timings
}
\author{
Sebastian Kreutzer
} 

\section{How to cite}{
Kreutzer, S. (2019). module_verify_ExampleTimings(): Veriyf Example Timings. In: Kreutzer, S., Burow, C. (2019). RLumBuild: RLum Universe Package BuildingR package version 0.1.0.9000-2. https://CRAN.R-project.org/package=RLumBuild
}

