% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_write_BibTeX.R
\name{module_write_BibTeX}
\alias{module_write_BibTeX}
\title{Create Package BibTeX file}
\usage{
module_write_BibTeX()
}
\description{
A BibTeX file to be used to cite the package is automatically created.
}
\section{Function version}{
 0.1.0
}

\author{
Sebastian Kreutzer, Institute of Geography, Heidelberg University (Germany)
, RLum Developer Team} 

\section{How to cite}{
Kreutzer, S., 2024. module_write_BibTeX(): Create Package BibTeX file. Function version 0.1.0. In: Kreutzer, S., Burow, C., 2024. RLumBuild: RLum Universe Package Building. R package version 0.1.8.9000-8. https://github.com/R-Lum/RLumBuild
}

