% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_write_BibTeX.R
\name{module_write_BibTeX}
\alias{module_write_BibTeX}
\title{Create Package BibTeX file}
\usage{
module_write_BibTeX()
}
\description{
A BibTeX file to be used to cite the package is automatically created.
}
\section{Function version}{
 0.1.0
}

\author{
Sebastian Kreutzer, IRAMAT-CRP2A, UMR 5060, CNRS - Université Bordeaux Montaigne (France)
, RLum Developer Team} 

\section{How to cite}{
Kreutzer, S., 2019. module_write_BibTeX(): Create Package BibTeX file. Function version 0.1.0. In: Kreutzer, S., Burow, C., 2019. RLumBuild: RLum Universe Package BuildingR package version 0.1.4.9000-16. https://github.com/R-Lum/RLumBuild
}

