% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_write_codemetar.R
\name{module_write_codemetar}
\alias{module_write_codemetar}
\title{Create Codemetar JSON-LD file}
\usage{
module_write_codemetar()
}
\description{
Add JSON-LD file file to make the package more tracable
}
\section{Function version}{
 0.1.0
}

\seealso{
[codemetar::create_codemeta]
}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, UMR 5060, CNRS - Université Bordeaux Montaigne (France)
, RLum Developer Team} 

\section{How to cite}{
Kreutzer, S., 2019. module_write_codemetar(): Create Codemetar JSON-LD file. Function version 0.1.0. In: Kreutzer, S., Burow, C., 2019. RLumBuild: RLum Universe Package BuildingR package version 0.1.4.9000-6. https://github.com/R-Lum/RLumBuild
}

