% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_write_codemetar.R
\name{module_write_codemetar}
\alias{module_write_codemetar}
\title{Create Codemetar JSON-LD file}
\usage{
module_write_codemetar()
}
\description{
Add JSON-LD file file to make the package more tracable
}
\section{Function version}{
 0.1.0
}

\seealso{
[codemetar::create_codemeta]
}
\author{
Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom)
, RLum Developer Team} 

\section{How to cite}{
Kreutzer, S., 2021. module_write_codemetar(): Create Codemetar JSON-LD file. Function version 0.1.0. In: Kreutzer, S., Burow, C., 2021. RLumBuild: RLum Universe Package Building. R package version 0.1.7.9000-1. https://github.com/R-Lum/RLumBuild
}

