% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_write_FunctionArgumentsList.R,
%   R/module_write_FunctionList.R
\name{module_write_FunctionList}
\alias{module_write_FunctionList}
\title{Write Function Argument List}
\usage{
module_write_FunctionList()

module_write_FunctionList()
}
\description{
Create a spread-sheet table with all function arguments

Create a HTML and markdown list of all major functions in the package
}
\details{
Pevious versions were written by Michael Dietze and Sebastian Kreutzer.
This version programmatically fetches all functions and arguments and does
not rely on parsing the .Rd files with regular expressions.
}
\section{Function version}{
 0.1.0

 0.1.0
}

\author{
Christoph Burow, Sebastian Kreutzer, IRAMAT-CRP2A, UMR 5060, CNRS - Université Bordeaux Montaigne (France)

Sebastian Kreutzer, IRAMAT-CRP2A, UMR 5060, CNRS - Université Bordeaux Montaigne (France)
} 

\section{How to cite}{
Burow, C., Kreutzer, S. (2019). module_write_FunctionList(): Write Function Argument List. Function version 0.1.0. In: Kreutzer, S., Burow, C. (2019). RLumBuild: RLum Universe Package BuildingR package version 0.1.1.9000-2. https://CRAN.R-project.org/package=RLumBuild
}

